### 2024-09-16 #7233 @editorialbot
[PRE REVIEW]: ExcelToShiny: An R Package for Automating Shiny Dashboard Creation from Excel Templates
**Submitting author:** @lilyclements (<a href="http://orcid.org/0000-0001-8864-0552">Lily Clements</a>)
**Repository:** https://github.com/IDEMSInternational/ExcelToShiny
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
