### 2024-09-16 #7233 @lilyclements
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/10884059299?check_suite_focus=true). Author (Lily Clements) is missing affiliation.

I believe I have now fixed this, apologies and thank you for flagging it.
