+++ DESCRIPTION
@@ -10,6 +10,7 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
 Imports:
+  desc,
   dplyr,
   forcats,
   ggplot2, 
