+++ DESCRIPTION
@@ -10,7 +10,6 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
 Imports:
-  desc,
   dplyr,
   forcats,
   ggplot2, 
