### 2024-10-08 33b8121
Adding functionality to have nrow
