### 2024-10-08 47e620a
fix bug in build_shiny
