### 2024-10-08 5c5efef
adding option to have data in the main_page
