### 2024-10-08 5288f8c
update to shiny vignette following additions
