### 2024-10-16 87c10a0
temp fix in plotly boxplot code as recorded in issue https://github.com/plotly/plotly.R/issues/2032
