### 2024-10-23 02be118
expand functionality to allow for more box types
