### 2024-10-23 #7233 @mbobra
👋 @crvernon Thank you for the invite and I'm definitely happy to edit a submission -- this one is not the right fit for me, though, because I don't know Excel or R. My apologies!
