### 2024-10-23 eb6aa26
added option to filter between dates with the addition of dateRangeInput; added new type for just data_frames
