+++ R/generate_from_ODK_functions.R
@@ -1,0 +1,164 @@
+#' Create a Data Frame for a Single "Box" Item
+#'
+#' `create_box_ODK` Helper function to create a single "box". Generates a single row of data representing a "box" with specified attributes.
+#'
+#' @param type Character. The type of the box. Default is "box".
+#' @param value Character. Specifies the box's value, often a summary type like "boxplot_summary".
+#' @param parameter_list Character. Parameters for the box, e.g., color or text, passed as a string.
+#' @param variable Character. Name of the variable represented in the box.
+#' @param row Integer. Row identifier for the box.
+#'
+#' @return A data frame with one row containing the specified box attributes.
+create_box_ODK <- function(type = "box", value = "boxplot_summary",
+                       parameter_list = NULL, variable = NULL,
+                       row = 1) {
+  return(data.frame(
+    type = type,
+    value = value,
+    parameter_list = parameter_list,
+    variable = variable,
+    row = row
+  ))
+}
+
+#' Generate Data Frame of Box Items from Metadata
+#'
+#' `create_from_type_ODK` processes metadata to create a data frame with "box" items for each row in the metadata, based on the type of data.
+#'
+#' @param metadata Data frame. Metadata to be processed, containing columns `type`, `label`, and `name`.
+#' @param value Character. Default value for box items, typically used for "boxplot_summary" or other summary types.
+#'
+#' @return A data frame containing rows of box items, each representing metadata attributes.
+create_from_type_ODK <- function(metadata = ODK_sample_survey_group, value = "boxplot_summary") {
+  if (nrow(metadata) == 0) return(NULL)
+  
+  metadata %>%
+    dplyr::mutate(
+      value = dplyr::case_when(
+        type %in% c("integer", "decimal") ~ "boxplot_summary",
+        TRUE ~ "bar_table"
+      ),
+      parameter_list = paste0("text = \"", label, "\", colour = \"blue\""),
+      variable = name,
+      row = row_number()
+    ) %>%
+    dplyr::rowwise() %>%
+    dplyr::summarise(
+      create_box(type = "box", value = value, parameter_list = parameter_list, variable = variable, row = row),
+      .groups = 'drop'
+    )
+}
+
+#' Generate Display Pages from Metadata Based on Group and Type
+#'
+#' `create_display_pages_from_metadata_ODK` filters metadata by group and type, then generates a display page as a data frame of box items.
+#'
+#' @param input_data Data frame. Input metadata, containing columns for `group`, `type`, `label`, and `name`.
+#' @param output_data Data frame. Placeholder for processed output (unused in this function).
+#' @param group_name Character. Name of the group to filter metadata by (e.g., "Identification").
+#'
+#' @return A data frame of display pages, each row representing a box item based on the specified metadata attributes.
+create_display_pages_from_metadata_ODK <- function(input_data = ODK_sample_survey, output_data, group_name = "Identification") {
+  # Filter based on type and group, then generate the display page
+  input_data <- input_data %>%
+    dplyr::filter(group == group_name, type %in% c("integer") | grepl("select_one|select_multiple", type))
+  
+  data_group <- create_from_type(input_data)
+  return(data_group)
+}
+
+#' Generate Data List for Display in Excel Template
+#'
+#' This function generates a data list structured as an Excel template, with grouped and miscellaneous variables displayed
+#' for creating tabs on a Shiny dashboard. It filters variables based on presence in both the form and dataset,
+#' organises them into specified groups, and adds miscellaneous items as a separate tab if needed.
+#'
[…]
