+++ NAMESPACE
@@ -4,6 +4,7 @@ export("%>%")
 export(PLH_shiny)
 export(add_na_variable)
 export(build_shiny)
+export(generate_data_list)
 export(get_parameter_value)
 export(naming_conventions)
 importFrom(magrittr,"%>%")
