### 2024-10-29 4faae59
fixing bug: if data is given in excel template and the cell is blank, it now takes the data_frame argument by default
