### 2024-10-30 213cee9
checks for contents$type and main_page$type is valid; checks that the value given in the filter data is valid, and adding automated checks; renaming contents1 to contents,
