### 2024-11-05 65bba88
fixes following update to R version 4.4.2
