+++ DESCRIPTION
@@ -8,7 +8,7 @@ Description: Provides functions for automating the creation of Shiny dashboards 
 License: LGPL (>= 3)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.2
 Imports:
   dplyr,
   forcats,
