### 2024-11-25 ece7eca
fixing bugs - issue if filter applied to the non-dominant data frame, allow for new lines in excel R code, allow for + or %>% to be missing or present in excel R code.
