+++ R/scatter_table.R
@@ -26,8 +26,12 @@ scatter_table <- function(data, variable, type = c("freq", "summary"), spreadshe
     # Command string from the spreadsheet
     command_string <- spreadsheet$data_manip
     
-    # Construct the full command
+    # Remove extra whitespace or control characters from command_string
+    command_string <- gsub("\r\n", "\n", command_string)  # Replace CRLF with LF for consistency
+    command_string <- gsub("^%>%\\s*", "", command_string) # Remove leading %>% if present
+    command_string <- paste0("data %>%", command_string)       # Append the dataset reference
     
+    # Construct the full command
     if (!is.null(grouped_vars)){
       full_command <- paste0("data %>% dplyr::group_by(", grouped_vars, ")", command_string)
     } else {
@@ -70,7 +74,14 @@ scatter_table <- function(data, variable, type = c("freq", "summary"), spreadshe
   
   plot_obj <- create_scatter_plot(data, variable)
   if (!is.null(spreadsheet$graph_manip) && !is.na(spreadsheet$graph_manip)) {
-    plot_command <- paste0("plot_obj + ", spreadsheet$graph_manip)
+    command_string <- spreadsheet$graph_manip
+    
+    # Clean and construct the full command
+    # Remove extra whitespace or control characters from command_string
+    command_string <- gsub("\r\n", "\n", command_string)  # Replace CRLF with LF for consistency
+    command_string <- gsub("^+\\s*", "", command_string) # Remove leading %>% if present
+    plot_command <- paste0("plot_obj + ", command_string)
+    
     plot_obj <- tryCatch({
       eval(parse(text = plot_command))
     }, error = function(e) {
