+++ R/bar_table.R
@@ -33,6 +33,12 @@ bar_table <- function(data, variable, type = c("freq", "summary"), spreadsheet, 
     # Command string from the spreadsheet
     command_string <- spreadsheet$data_manip
     
+    # Clean and construct the full command
+    # Remove extra whitespace or control characters from command_string
+    command_string <- gsub("\r\n", "\n", command_string)  # Replace CRLF with LF for consistency
+    command_string <- gsub("^%>%\\s*", "", command_string) # Remove leading + if present
+    command_string <- paste0("data %>%", command_string)       # Append the dataset reference
+    
     # Construct the full command
     if (!is.null(grouped_vars)){
       full_command <- paste0("data %>% dplyr::group_by(", grouped_vars, ")", command_string)
@@ -86,7 +92,13 @@ bar_table <- function(data, variable, type = c("freq", "summary"), spreadsheet, 
 
   plot_obj <- create_histogram_plot(data, variable, grouped_vars)
   if (!is.null(spreadsheet$graph_manip) && !is.na(spreadsheet$graph_manip)) {
-    plot_command <- paste0("plot_obj + ", spreadsheet$graph_manip)
+    command_string <- spreadsheet$graph_manip
+    
+    # Clean and construct the full command
+    # Remove extra whitespace or control characters from command_string
+    command_string <- gsub("\r\n", "\n", command_string)  # Replace CRLF with LF for consistency
+    command_string <- gsub("^+\\s*", "", command_string) # Remove leading %>% if present
+    plot_command <- paste0("plot_obj + ", command_string)
     
     plot_obj <- tryCatch({
       eval(parse(text = plot_command))

[…]
