### 2024-11-25 49c9306
fixes in boxplot_table
