### 2024-11-25 514ddc9
fixing issue with icon NA
