+++ R/display_sheet.R
@@ -35,26 +35,33 @@ display_sheet <- function(data_list, spreadsheet_name, d_box, status = "primary"
     }
     tab_item_objects[[l]] <- split_layout(row_l_set)
   }
-
+  
   for (l in 1:length(tab_item_objects)){
     tab_item_objects[[l]] <- shiny::fluidRow(shiny::column(12,
                                                            align = "center",
                                                            tab_item_objects[[l]]),
                                              width = 10)
   }
-
+  
   main_page_info <- which(data_list[["contents"]][["ID"]] == spreadsheet_name)
   main_page_info <- data_list[["contents"]][main_page_info,]
+  
+  if (is.null(main_page_info$icon) || (!is.null(main_page_info$icon) & is.na(main_page_info$icon))){
+    shiny_layout_display <- shiny::h2(main_page_info$name)                                                                
+  } else {
+    shiny_layout_display <- shiny::splitLayout(shiny::h2(main_page_info$name), 
+                                               shiny::icon(main_page_info$icon, "fa-6x"),
+                                               cellArgs = list(style = "vertical-align: top"), 
+                                               cellWidths = c("80%", "20%"))
+  }
+  
   
   tab_item <- shinydashboard::tabItem(tabName = data_list$contents$ID[[j]],
                                       
                                       # Stuff for the top of the tab
                                       shiny::fluidRow(shiny::column(12,
                                                                     align = "center",
-                                                                    shinydashboard::box(shiny::splitLayout(shiny::h2(main_page_info$name), 
-                                                                                                           shiny::icon(main_page_info$icon, "fa-6x"),
-                                                                                                           cellArgs = list(style = "vertical-align: top"), 
-                                                                                                           cellWidths = c("80%", "20%")),
+                                                                    shinydashboard::box(shiny_layout_display,
                                                                                         status = status,
                                                                                         background = colour,
                                                                                         width = 10,
