+++ vignettes/Excel_Sheet_Explained.Rmd
@@ -2,8 +2,8 @@
 ---
 title: "Using Excel for Shiny Authoring System"
 author: "Lily Clements"
-date: "r Sys.Date()"
 output:
+  pdf_document: default
   html_document:
     df_print: paged
 vignette: "%\\VignetteIndexEntry{Using Excel for Shiny Authoring System} %\\VignetteEngine{knitr::rmarkdown}
@@ -38,11 +38,11 @@ When creating a spreadsheet for the Shiny authoring system, there are five key a
 4. **Tabbed Display Page**: Manages tabbed display sections.
 5. **Download Page**: Handles downloadable content and related settings.
 
-You can view a sample Excel spreadsheet on GitHub titled "nhanes_data.xlsx". This can be found under vignettes/data/.
+You can view a sample Excel spreadsheet on GitHub titled `"nhanes_data.xlsx"`. This can be found under `vignettes/data/`.
 
 ### Example Code
 
-Below is the R code needed to test the provided Excel spreadsheet and see how it integrates with the PLHShiny package:
+Below is the R code needed to test the provided Excel spreadsheet and see how it integrates with the `ExcelToShiny` package:
 
 ```{r eval = FALSE}
 library(rio)
@@ -159,7 +159,7 @@ Here we have three different components: value boxes, filtering, and group boxes
   - filter_box: Provides filtering options.
   - group_by_box: Provides grouping options.
 - **Value**: Takes different values at the moment depending on what type is.
-  - If type is value_box, then the options here are value_box_rows (which gives the number of rows), value_box (which gives the frequency of rows for the variable given in the variable parameter), mean_box (which gives the mean for the variable given in the variable parameter), and mean_sd_box (which gives the mean and SD for the variable given in the variable parameter)
+  - If type is value_box, then the options here are value_box_rows (which gives the number of rows), value_box (which gives the frequency of rows for the variable given in the variable parameter), mean_box (which gives the mean for the variable given in the variable parameter), and mean_sd_box (which gives the mean and SD for the variable given in the variable parameter). You can also specify any function valid in R, such as `sum_box`, `median_box`, and so forth.
   - If type is filter_box, then the options are "checkbox_group" and "dates". The former gives a set of checkbox options, and the latter can be used to filter by dates.
   - If type is group_by_box, then the only options currently is "checkbox_group", which gives a set of checkbox options.
 - **Parameter List**: Parameters that the item declared in type takes. E.g., for value_box, parameters include "text", "colour", and "icon". For filter_box, parameters are "label", "choices", and "selected".
@@ -169,7 +169,7 @@ Here we have three different components: value boxes, filtering, and group boxes
   - _Example_: If displaying data for "gender == 'males'", then "males" should be specified here.
 
 | **type**     | **value**      | **parameter_list**                                                                                                     | **variable_value**                                             | **data**
-|--------------|----------------|-----------------------------------------------------------------------------------------------------------------------|---------------------------------------------------------------|----------------------------------------|
+|---------------------------|------------------------------|------------------------------------------------------------------------------------------------------------------|----------------------------------------------------------|------------------------------------|
 | value_box  | value_box_rows    | Accepted values are: "text", "colour", "icon". These read into valueBox in Shiny.                             | Not relevant. | Name of data frame to use. |
 |              |                | - "text": The value for the "subtitle" parameter. This is the text displayed in the value box (e.g., "Males in the study"). |                                                               | |
[…]
