+++ R/PLH_shiny.R
@@ -1,6 +1,5 @@
 #' PLH Shiny Function - Create and Launch a Shiny Dashboard
-#' 
-#' This function is now Depreciated. Look at `build_shiny` instead.
+#'
 #' The PLH_shiny function generates a Shiny dashboard application based on the provided datasets and specifications. This function is designed to create a dynamic and interactive user interface that includes value boxes, filters, groupings, plots, tables, and downloadable content, all customisable through the input parameters.
 #'
 #' @param title A string specifying the title of the dashboard.
@@ -64,6 +63,19 @@ PLH_shiny <- function (title, data_list, data_frame, status = "primary", colour 
   }
   
   # Setting up (pre-UI and pre-server items) --------------------------------
+  # Check the types in contents are all valid types (display, tabbed_display, and download)
+  data_list$contents <- data_list$contents %>%
+    mutate(type = ifelse(stringdist::stringdist(type, "Display", method = "lv") <= 2, "Display",
+                         ifelse(stringdist::stringdist(type, "Tabbed_display", method = "lv") <= 3, "Tabbed_display",
+                                ifelse(stringdist::stringdist(type, "Download", method = "lv") <= 3, "Download",
+                                       type))))
+  valid_contents_type <- c("Display", "Tabbed_display", "Download")
+  if (!all(data_list$contents$type %in% valid_contents_type)){
+    invalid_contents_type <- data_list$contents %>%
+      dplyr::filter(!type %in% valid_contents_type) %>%
+      dplyr::pull(type)
+    stop("Cannot read contents type: ", paste0(invalid_contents_type, sep = ", "), "Should be one of ", paste0(valid_contents_type, sep = ", "))
+  }
   contents <- data_list$contents
   
   for (i in 1:length(data_list)){
@@ -72,26 +84,39 @@ PLH_shiny <- function (title, data_list, data_frame, status = "primary", colour 
     }
   }
   
-  
   # list of sheets
-  
   # check the variables exist
-  for (df_name in names(data_l)){
+  for (df_name in names(data_list)){
     sheet <- data_list[[df_name]]
     data_frame_name <- deparse(substitute(data_frame))
     results <- NULL
     results <- check_variables_existence(sheet, data_frame = data_frame_name)
     if (!is.null(results) && nrow(results) > 0) check_unknown_variables(results)
   }
-
+  
   # Populate items for the tabs ------------------------------------------------
-  display_box <- display_contents(data_frame = data_frame, contents1 = contents, data_list = data_list, k = which(data_list$contents$type == "Tabbed_display"))
+  display_box <- display_contents(data_frame = data_frame, contents = contents, data_list = data_list, k = which(data_list$contents$type == "Tabbed_display"))
+  
   my_tab_items <- create_tab_items(data_list = data_list,
                                    d_box = display_box,
                                    status = status,
                                    colour = colour)
   
-  # value box for main page ----------------------------------------------------
+  # value box for main page -------------------------------------------------------------------------------
+  # check type is one of value_box, filter_box, or group_by_box
+  data_list$main_page <- data_list$main_page %>%
+    mutate(type = ifelse(stringdist::stringdist(type, "value_box", method = "lv") <= 2, "value_box",
+                         ifelse(stringdist::stringdist(type, "filter_box", method = "lv") <= 3, "filter_box",
+                                ifelse(stringdist::stringdist(type, "group_by_box", method = "lv") <= 3, "group_by_box",
+                                       type))))
+  valid_boxes <- c("value_box", "filter_box", "group_by_box")
+  if (!all(data_list$main_page$type %in% valid_boxes)){
+    invalid_type <- data_list$main_page %>%
+      dplyr::filter(!type %in% valid_boxes) %>%
+      dplyr::pull(type)
[…]
