### 2024-12-11 795de72
fixes in icon and table display for specify_table
