+++ R/specify_table.R
@@ -19,9 +19,14 @@ specify_table <- function(data, spreadsheet, grouped_vars = NULL) {
     return(all_return)
   }
   
-  #if (!is.null(grouped_vars)) grouped_vars <- NULL
+  # TODO: need to fix this up to work for grouped_vars.
+  # but if I have facets, it removes them. I think (see facilitator mexico, with the df having grouped var)
+  if (!is.null(grouped_vars)) grouped_vars <- NULL
   
   if (!is.null(grouped_vars)){
+    if (!all(grouped_vars %in% names(data))) {
+      grouped_vars <- grouped_vars[which(grouped_vars %in% names(data))]
+    }
     group_cmd <- paste0("%>% group_by(", grouped_vars, ")")
   } else {
     group_cmd <- paste0("")
@@ -35,22 +40,13 @@ specify_table <- function(data, spreadsheet, grouped_vars = NULL) {
     # Clean and construct the full command
     # Remove extra whitespace or control characters from command_string
     command_string <- gsub("\r\n", "\n", command_string)   # Replace CRLF with LF for consistency
-    command_string <- gsub("^%>%\\s*", "", command_string) # Remove leading %>% if present
-    
-    # Construct the full command
-    if (!is.null(grouped_vars)){
-      full_command <- paste0("data %>% dplyr::group_by(", grouped_vars, ") %>% ", command_string)
-    } else {
-      full_command <- paste0("data %>% ", command_string)
+    #command_string <- gsub("^%>%\\s*", "", command_string) # Remove leading %>% if present
+
+    if (startsWith(trimws(spreadsheet$table_manip), "%>%")){
+      all_return$table <- eval(parse(text = paste0("data ", group_cmd, command_string)))
+    } else{
+      all_return$table <- eval(parse(text = command_string))      
     }
-    
-    full_command <- tryCatch({
-      eval(parse(text = full_command))
-    }, error = function(e) {
-      message("Ignoring manipulations. Error in evaluating data manipulation command: ", e$message)
-      data  # Return NULL or handle the error as appropriate
-    })
-    all_return$table <- full_command 
   } else {
     if (!is.null(spreadsheet$data_manip) && !is.na(spreadsheet$data_manip)) {
       warning("Manipulations for specify_table are given in data_manip. These should be in table_manip.")
@@ -81,5 +77,8 @@ specify_table <- function(data, spreadsheet, grouped_vars = NULL) {
     }
   }
   all_return$plot_obj <- NULL
+  
+  #print(all_return$table)
+  
   return(all_return)
 }
