### 2025-03-05 #7876 @editorialbot
[REVIEW]: ExcelToShiny: An R Package for Automating Shiny Dashboard Creation from Excel Templates
**Submitting author:** @lilyclements (<a href="http://orcid.org/0000-0001-8864-0552">Lily Clements</a>)
**Repository:** https://github.com/IDEMSInternational/ExcelToShiny
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.0
