### 2025-03-19 #7876 @skanwal
@devarops the automated instructions for starting the review process can be found in the [first comment](https://github.com/openjournals/joss-reviews/issues/7876#issue-2898598607) by bot on this issue.

Please let me know if you have any questions.
