### 2025-03-20 IDEMSInternational/ExcelToShiny#12 @devarops
[JOSS REVIEW] Fix the R-CMD-Check Workflow
The [R-CMD-check workflow](https://github.com/IDEMSInternational/ExcelToShiny/actions/workflows/check-release.yaml) is currently failing. Could you please address the errors to ensure it passes?

---

This issue is part of the review process for a submission to the Journal of Open-Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
