### 2025-03-20 IDEMSInternational/ExcelToShiny#13 @devarops
[JOSS REVIEW] Fix the Test Coverage Workflow
The [test-coverage workflow](https://github.com/IDEMSInternational/ExcelToShiny/actions/workflows/test-coverage.yaml) is not generating results. Additionally, the [Codecov report](https://app.codecov.io/github/IDEMSInternational/ExcelToShiny) indicates 0% test coverage.
Could you please investigate and resolve these issues to ensure the coverage report populates accurately?

---

This issue is part of the review process for a submission to the Journal of Open-Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
