### 2025-03-20 IDEMSInternational/ExcelToShiny#14 @devarops
[JOSS REVIEW] Add Community Contribution Guidelines
Please add community guidelines to clarify processes for:

1. **Contributing to the software** (e.g., pull request standards, coding conventions).
1. **Reporting issues** (e.g., templates for bug reports or feature requests).
1. **Seeking support** (e.g., communication channels, expected response times).

These guidelines will help foster a collaborative environment and streamline interactions with contributors.

---

This issue is part of the review process for a submission to the Journal of Open-Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
