### 2025-03-20 IDEMSInternational/ExcelToShiny#15 @devarops
[JOSS REVIEW] Compare ExcelToShiny to Commonly Used Packages
Could you please add a **"State of the Field"** section describing how `ExcelToShiny` compares to other commonly used packages, such as `flexdashboard` and `golem`? This would help users understand its unique value proposition within the existing ecosystem.

---

This issue is part of the review process for a submission to the Journal of Open-Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
