### 2025-03-20 IDEMSInternational/ExcelToShiny#16 @devarops
[JOSS REVIEW] Add a "References" Section Heading at the End of the Paper
The list of references currently appears immediately after the Conclusion without a section heading. Could you please add a **"References"** heading at the end of the paper to create a dedicated section for the references?

---

This issue is part of the review process for a submission to the Journal of Open-Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
