### 2025-03-20 IDEMSInternational/ExcelToShiny#17 @devarops
[JOSS REVIEW] Add URL to Power BI Reference
The Power BI reference entry currently does not include a URL. Could you please add a link to this entry?

https://github.com/IDEMSInternational/ExcelToShiny/blob/6d581cf4bb51b95017b113ff43053158974de714/joss_papers/paper.bib#L19-L24

---

This issue is part of the review process for a submission to the Journal of Open-Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
