### 2025-03-20 IDEMSInternational/ExcelToShiny#18 @devarops
[JOSS REVIEW] Update README to Reference `ExcelToShiny` Instead of `plhR`
The README file currently refers to `plhR`, likely due to a repository rename or outdated template. Could you please update all instances of `plhR` to `ExcelToShiny` in the README file to ensure consistency with the current project name?

---

This issue is part of the review process for a submission to the Journal of Open-Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
