### 2025-03-20 #7876 @devarops
Hello @lilyclements, 👋🏾

Congratulations on `ExcelToShiny`! 🎉
It’s a fantastic tool for non-technical users who want to create Shiny dashboards with ease.

Below are my review comments for your reference:

- 📋 **Review feedback**: [Link to GitHub issues](https://github.com/IDEMSInternational/ExcelToShiny/issues?q=author%253Adevarops)

Thank you for your hard work, and let me know if you need further clarification!
