+++ README.Rmd
@@ -13,12 +13,12 @@ knitr::opts_chunk$set(
 )
 ```
 
-# plhR
+# ExcelToShiny
 
 <!-- badges: start -->
-[![R-CMD-check](https://github.com/IDEMSInternational/plhR/workflows/R-CMD-check/badge.svg)](https://github.com/IDEMSInternational/plhR/actions)
+[![R-CMD-check](https://github.com/IDEMSInternational/ExcelToShiny/workflows/R-CMD-check/badge.svg)](https://github.com/IDEMSInternational/ExcelToShiny/actions)
 [![Codecov test
-coverage](https://codecov.io/gh/IDEMSInternational/plhR/branch/main/graph/badge.svg)](https://app.codecov.io/gh/IDEMSInternational/plhR?branch=main)
+coverage](https://codecov.io/gh/IDEMSInternational/ExcelToShiny/branch/main/graph/badge.svg)](https://app.codecov.io/gh/IDEMSInternational/ExcelToShiny?branch=main)
 [![Lifecycle:
 experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental)
 [![Project Status: WIP – Initial development is in progress, but there
@@ -31,7 +31,7 @@ A set of functions useful in analysing PLH data.
 
 ## Installation
 
-You can install the development version of plhR from [GitHub](https://github.com/) with:
+You can install the development version of ExcelToShiny from [GitHub](https://github.com/) with:
 
 ```{r}
 # install.packages("devtools")
