### 2025-03-20 9fe57d7
Updating README file and fixing checks
