+++ DESCRIPTION
@@ -23,6 +23,7 @@ Imports:
   shinyauthr,
   shinydashboard,
   shinyjs,
+  stringdist,
   tibble,
   tidyr,
   viridis
