### 2025-03-20 IDEMSInternational/ExcelToShiny#13 @lilyclements
@devarops thank you for this. I have not yet implemented testing in this, but I agree this should be a priority. Can you remind me if JOSS requires test coverage? If not, then I will do a few tests each evening or so, but if they do then I will prioritise it above other projects I am working on.
