### 2025-03-20 3a5cbf1
Adding State of the Field section
