+++ joss_papers/paper.md
@@ -20,7 +20,7 @@ authors:
 affiliations:
  - name: IDEMS International
    index: 1
-date: 13th September
+date: 20th March 2025
 bibliography: paper.bib
 
 ---
