+++ joss_papers/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-The `ExcelToShiny` package is an open-source, user-friendly tool that automates the creation of Shiny dashboards from structured Excel templates. Designed to simplify the process of building interactive web applications in R, `ExcelToShiny` allows users to convert Excel spreadsheets directly into customisable, fully functional dashboards with minimal coding. This solution promotes reproducibility and transparency while significantly reducing the technical barrier for users seeking data-driven insights through visualisations.
+`ExcelToShiny` is an open-source, user-friendly tool that automates the creation of Shiny dashboards from structured Excel templates. Designed to simplify the process of building interactive web applications in R, `ExcelToShiny` allows users to convert Excel spreadsheets directly into customisable, fully functional dashboards with minimal coding. This solution promotes reproducibility and transparency while significantly reducing the technical barrier for users seeking data-driven insights through visualisations.
 
 By automating the traditionally labour-intensive process of dashboard generation, `ExcelToShiny` ensures that even non-technical users can engage with dynamic data visualisations. The package plays a role in promoting accessibility and flexibility in data presentation, enabling users to focus on decision-making rather than complex programming tasks.
 
@@ -37,7 +37,7 @@ However, constructing these dashboards from scratch in Shiny can be complex even
 
 Researchers can face difficulties in creating reproducible workflows for interactive data visualisations, especially when technical expertise is a barrier. Static reports limit data exploration [@koldus2024dashboards], and proprietary software tools, while convenient, are not as easily adaptable or transparent enough for scientific research needs [@biswas2021tableau]
 
-`ExcelToShiny` addresses these gaps by offering an open-source, R-based alternative that automates the conversion of Excel templates into Shiny dashboards. 
+`ExcelToShiny` addresses these gaps by offering an R-based alternative that automates the conversion of Excel templates into Shiny dashboards. 
 By reading a user-defined Excel structure, `ExcelToShiny` dynamically generates the corresponding UI and server components for a Shiny application. This approach allows users to create dashboards that are reproducible, transparent, and easy to share, with the flexibility to modify and adapt them as needed within the R ecosystem. 
 Additionally, by lowering the technical barrier, `ExcelToShiny` allows researchers from various fields to engage with large datasets, perform real-time analysis, and generate data-driven insights without the need for extensive coding.
 
@@ -51,7 +51,7 @@ While these tools are powerful, they introduce complexities that may not suit al
 
 # Functionality and Features
 
-`ExcelToShiny` leverages a simple, structured Excel format to automatically generate Shiny dashboards, making the process highly accessible. The package interprets Excel template elements like tabs, graphs, tables, and menus, converting them into Shiny UI components and server logic with minimal user input. The dashboards can have regular display pages, display pages with tabs, or have a page that downloads the data. Furthermore, while much of the process is automated, users can further customise their dashboards, enabling fine-tuning of both layout and behavior to meet specific requirements. This allows  users to fine-tune both the appearance and functionality without needing to dive into R code.
+`ExcelToShiny` leverages a simple, structured Excel format to automatically generate Shiny dashboards, making the process highly accessible. The package interprets Excel template elements like tabs, graphs, tables, and menus, converting them into Shiny UI components and server logic with minimal user input. The dashboards can have regular display pages, display pages with tabs, or have a page that downloads the data. Furthermore, while much of the process is automated, users can further customise their dashboards, enabling fine-tuning of both layout and behaviour to meet specific requirements without needing to dive into R code.
 
[…]
