### 2025-03-20 IDEMSInternational/ExcelToShiny#15 @lilyclements
Many thanks for this insight @devarops. I hope that I have now addressed this following the merge of PR #22. Let me know if you are happy with the changes and if this issue can be closed.
