### 2025-03-20 a0fb13c
fix in get_variables_existence for testing purposes
