### 2025-03-20 9c31817
build_shiny test fix
