### 2025-03-20 IDEMSInternational/ExcelToShiny#14 @devarops
I noticed that you suggest seeking support by opening a discussion in GitHub Discussions. However, it appears that GitHub Discussions is not currently activated in your repository.

Could you please either enable GitHub Discussions under your repository settings or update the documentation to reflect an alternative support method?
This would help ensure users can follow the instructions smoothly.

Thank you!
