### 2025-03-21 e1e9b9a
shiny tests to run app
