+++ tests/testthat/tests-build_shiny.R
@@ -62,7 +62,6 @@ test_that("create_shiny_dashboard runs successfully", {
       library(plotly)
       library(rio)
       library(NHANES)
-      print(project_dir)
       devtools::load_all()
       
       # Load data inside background session too!
