### 2025-03-21 9407706
test: build_shiny launches app correctly
