### 2025-03-21 ff92219
wd setup
