### 2025-03-21 3fff4bb
separating server into functions
