+++ R/server_functions.R
@@ -1,0 +1,336 @@
+# Helper Function 1: Prepare the data_list and extract df names
+prepare_data_list <- function(data_list, data_frame_name) {
+  for (i in seq_along(data_list)) {
+    if ("data" %in% names(data_list[[i]])) {
+      data_list[[i]] <- data_list[[i]] %>%
+        dplyr::mutate(data = ifelse(is.na(data), data_frame_name, data))
+    }
+  }
+  
+  df_names <- purrr::map(data_list, ~ .x$data) %>% unlist() %>% unique()
+  df_names <- df_names[!is.na(df_names)]
+  
+  return(list(data_list = data_list, list_of_df_names = df_names))
+}
+
+# Helper Function 2: Generate complete_dfs environment
+generate_complete_dfs <- function(df_names, env = parent.frame()) {
+  complete_dfs <- list()
+  for (df_name in df_names) {
+    new_name <- paste0(df_name, "_1")
+    stored_data <- get(df_name, envir = env)
+    assign(new_name, stored_data, envir = env)
+    complete_dfs[[new_name]] <- stored_data
+  }
+  complete_dfs
+}
+
+# Helper Function 3: Create reactive data filtering
+create_reactive_expression <- function(df_name, complete_dfs, key_var, valid_ids, grouped_vars = NULL) {
+  force(df_name)
+  shiny::reactive({
+    filtered_data_frame <- complete_dfs[[paste0(df_name, "_1")]]
+    
+    if (!is.null(grouped_vars)) {
+      filtered_data_frame <- filtered_data_frame %>%
+        dplyr::full_join(grouped_vars)
+    }
+    
+    if (!is.null(key_var) && key_var %in% names(filtered_data_frame)) {
+      filtered_data_frame <- filtered_data_frame %>% 
+        dplyr::filter(.data[[key_var]] %in% valid_ids())
+    }
+    
+    return(filtered_data_frame)
+  })
+}
+
+# Helper Function 4: Display content by tab
+display_content_by_tab <- function(tab_name, input, filtered_data, contents, data_list, list_of_reactives) {
+  if (input$tab == tab_name) {
+    server_display_contents(
+      data_frame = filtered_data(),
+      contents = contents,
+      data_list = data_list,
+      k = which(data_list$contents$type == "Tabbed_display"),
+      id_name = tab_name,
+      list_of_reactives = list_of_reactives
+    )
+  } else {
+    NULL
+  }
+}
+
+# Helper Function 5: Process spreadsheet for value boxes
+process_spreadsheet_function <- function(spreadsheet) {
+  spreadsheet_df <- spreadsheet %>%
+    dplyr::filter(type == "value_box") %>%
+    dplyr::select(name, parameter_list) %>%
+    tidyr::separate_rows(parameter_list, sep = ",\\s*") %>%
+    tidyr::separate(parameter_list, into = c("names", "values"), sep = "\\s*=\\s*") %>%
+    dplyr::mutate(values = stringr::str_remove_all(values, stringr::fixed("\"")))
+  return(spreadsheet_df)
+}
+
+# Helper Function 6: Render value boxes
+draw_top_value_boxes <- function(spreadsheet_shiny_value_box, processed_spreadsheet_data, filtered_data, list_of_reactives, output) {
+  shiny::observe({
+    lapply(seq_len(length(unique(spreadsheet_shiny_value_box$name))), function(i) {
+      ID <- spreadsheet_shiny_value_box[i,]$name
+      top_box <- top_value_boxes(
+        data_frame = filtered_data(),
+        spreadsheet = spreadsheet_shiny_value_box,
+        processed_spreadsheet = processed_spreadsheet_data,
+        unique_ID = ID,
[…]
