### 2025-03-21 f808b9e
tests to build shiny
