+++ .github/workflows/test-coverage.yaml
@@ -1,28 +1,19 @@
+# Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
+# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
     branches: [main, master]
   pull_request:
     branches: [main, master]
-
 name: test-coverage.yaml
-
 permissions: read-all
-
 jobs:
   test-coverage:
     runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-
     steps:
       - uses: actions/checkout@v4
-
-      - name: Install dependencies for shinytest2 / chromote
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y libnss3 libatk-bridge2.0-0 libgtk-3-0 libdrm2 libgbm1 libasound2
-          sudo apt-get install -y chromium-browser
-          echo "CHROMOTE_CHROME_PATH=/usr/bin/chromium-browser" >> $GITHUB_ENV
 
       - uses: r-lib/actions/setup-r@v2
         with:
@@ -30,36 +21,24 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: |
-            any::covr
-            any::xml2
-            any::rio
-            any::NHANES
-            any::callr
-            any::shiny
-            any::shinydashboard
-            any::stringdist
-            any::plotly
-            any::shinytest2
-
-      - name: Run tests and collect coverage
+          extra-packages: any::covr, any::xml2
+          needs: coverage
+      - name: Test coverage
         run: |
-          cov <- covr::package_coverage(quiet = FALSE)
+          cov <- covr::package_coverage(
+            quiet = FALSE,
+            clean = FALSE,
+            install_path = file.path(normalizePath(Sys.getenv("RUNNER_TEMP"), winslash = "/"), "package")
+          )
           covr::to_cobertura(cov)
         shell: Rscript {0}
 
-      - uses: codecov/codecov-action@v4
-        with:
-          fail_ci_if_error: ${{ github.event_name != 'pull_request' && true || false }}
-          file: ./cobertura.xml
-          plugin: noop
-          disable_search: true
-          token: ${{ secrets.CODECOV_TOKEN }}
-
+	@@ -49,13 +59,12 @@ jobs:
       - name: Show testthat output
         if: always()
         run: |
-          find '${{ runner.temp }}' -name 'testthat.Rout*' -exec cat '{}' \; || true
+          ## --------------------------------------------------------------------
+          find '${{ runner.temp }}/package' -name 'testthat.Rout*' -exec cat '{}' \; || true
         shell: bash
 
       - name: Upload test results
@@ -67,4 +46,4 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: coverage-test-failures
-          path: ${{ runner.temp }}
+          path: ${{ runner.temp }}/package
