### 2025-03-21 838b983
tests on helper functions
