### 2025-03-21 IDEMSInternational/ExcelToShiny#31 @devarops
[JOSS REVIEW] Update the About blurb
I noticed that the description in the About section of the repository currently states:

> A set of functions useful in PLH analysis

![Image](https://github.com/user-attachments/assets/2355c6d0-5404-4cb0-b30b-474cf0ee7773)

Could you please update the About section to reflect the current description of the repository?

---

This issue is part of the review process for a submission to the Journal of Open Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
