### 2025-03-21 559d1af
adding helper function tests
