### 2025-03-21 b66d2b1
update in helper test functions
