### 2025-03-25 29d4713
tests on tables/graphs
