+++ DESCRIPTION
@@ -11,6 +11,7 @@ Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
 Imports:
   dplyr,
+  DT,
   forcats,
   ggplot2, 
   gt, 
