### 2025-03-25 5939c90
adding tests for main_page_filter
