### 2025-03-25 a6b4527
further tests for boxplot_table
