### 2025-03-25 0b0a857
updates to plot and tbale tests
