### 2025-03-25 0f82eb3
update in plot table tests
