### 2025-03-26 9365c32
Adding tests for filter_main_page
