### 2025-03-26 8194871
tests in error handling plots
