### 2025-03-26 5814f12
Tests and fixes in R code
