+++ tests/testthat/tests-build_shiny.R
@@ -268,3 +268,68 @@ test_that("create_shiny_dashboard runs successfully for filter by date_range", {
                      deploy_shiny = FALSE)
   expect_equal(class(app), "list")
 })
+
+
+test_that("build_shiny runs successfully for incorrect values", {
+  library(ggplot2)
+  data(economics)
+  economics_data_list <- rio::import_list("testdata/economics_data.xlsx")
+  economics_data_list$main_page <- economics_data_list$main_page[1:2,]
+  
+  economics_data_list1 <- economics_data_list
+  economics_data_list1$demographics$parameter_list[1] <- "text = \"Personal Consumption Expenditure (in billions of dollars)\""
+  expect_warning(build_shiny(title = "New Dashboard", 
+                                 data_list = economics_data_list1,
+                                 data_fram = economics,
+                                 status = "primary",
+                                 colour = "lime",
+                                 key_var = "date",
+                                 deploy_shiny = FALSE),
+                 "Valid colours are: blue, green, light blue, orange, red. Setting colour to 'blue'.")
+  
+  economics_data_list_1 <- economics_data_list
+  economics_data_list_1$contents$type <- "none"
+  expect_error(build_shiny(title = "New Dashboard", 
+                                 data_list = economics_data_list_1,
+                                 data_fram = economics,
+                                 status = "primary",
+                                 colour = "blue",
+                                 key_var = "date",
+                                 deploy_shiny = FALSE),
+               "Cannot read contents type: none, Should be one of Display, Tabbed_display, Download,")
+
+  economics_data_list_1 <- economics_data_list
+  economics_data_list_1$main_page$type[1] <- "unknown_box"
+  expect_error(build_shiny(title = "New Dashboard", 
+                           data_list = economics_data_list_1,
+                           data_fram = economics,
+                           status = "primary",
+                           colour = "blue",
+                           key_var = "date",
+                           deploy_shiny = FALSE),
+               "Cannot read type: unknown_box, on main_page. Should be one of value_box, filter_box, group_by_box,")
+  
+  economics_data_list_1 <- economics_data_list
+  economics_data_list_1$main_page <- rbind(economics_data_list_1$main_page, economics_data_list_1$main_page, economics_data_list_1$main_page)
+  app <- build_shiny(title = "New Dashboard", 
+                     data_list = economics_data_list_1,
+                     data_fram = economics,
+                     status = "primary",
+                     colour = "blue",
+                     key_var = "date",
+                     deploy_shiny = FALSE)
+  expect_equal(class(app), "list")
+  
+  # runs when there is no main_page
+  economics_data_list_1 <- economics_data_list
+  economics_data_list_1$main_page <- NULL
+  app <- build_shiny(title = "New Dashboard", 
+                     data_list = economics_data_list_1,
+                     data_fram = economics,
+                     status = "primary",
+                     colour = "blue",
+                     key_var = "date",
+                     deploy_shiny = FALSE)
+  expect_equal(class(app), "list")
+})
+
