### 2025-03-26 f0c6e4d
tests for server_display_* functions
