### 2025-03-26 08499fd
Test with summary_calculation function
