### 2025-03-26 dd9b58c
tests in summary_table
