### 2025-03-26 a24ee3b
Adding tests to further refine
