+++ R/get_parameter_value.R
@@ -33,56 +33,36 @@
 #' # Parameter not found
 #' param_not_found <- 'other_param = 42'
 #' get_parameter_value(param_not_found)  # Returns NULL
-get_parameter_value <- function(spreadsheet_parameters, name = "label", list = FALSE, logical = FALSE, date = FALSE){
-  #spreadsheet_parameters <- trimws(spreadsheet_parameters)
-  if (stringr::str_detect(spreadsheet_parameters, paste0(name, " ="))) {
-    if (stringr::str_detect(spreadsheet_parameters, paste0(name, " = "))) {
-      label_form = paste0(name, " = ")
-    } else {
-      label_form = paste0(name, " =")
-    }
-  } else if (stringr::str_detect(spreadsheet_parameters, paste0(name, "="))){
-    if (stringr::str_detect(spreadsheet_parameters, paste0(name, "= "))) {
-      label_form = paste0(name, "= ")
-    } else {
-      label_form = paste0(name, "=")
-    }
-  } else {
-    label_form = NULL
+get_parameter_value <- function(spreadsheet_parameters, name = "label", list = FALSE, logical = FALSE, date = FALSE) {
+  param_string <- spreadsheet_parameters
+  
+  # Pattern: name = value (with optional spaces)
+  if (date) {
+    # Match date as string or wrapped in as.Date()
+    date_pattern <- paste0(name, "\\s*=\\s*(?:as.Date\\()?[\"'](.*?)[\"']\\)?")
+    match <- stringr::str_match(param_string, date_pattern)[, 2]
+    return(if (!is.na(match)) match else NULL)
   }
   
-  if (!is.null(label_form)){
-    if (date == TRUE){
-      label_pattern <- paste0(label_form, '"(.*?)"')  # Match as.Date("value")
-      label_match <- regmatches(spreadsheet_parameters, regexec(label_pattern, spreadsheet_parameters))[[1]][1]
-      
-      # if it's NA, check in case as.Date has been added:
-      if (is.na(label_match)){
-        # Match for both simple and as.Date formats
-        label_form <- paste0(label_form, "as.Date\\(")  # Adjust for as.Date() structure
-        label_pattern <- paste0(label_form, '"(.*?)"\\)')  # Match as.Date("value")
-        
-        # Extract the full match
-        label_match <- regmatches(spreadsheet_parameters, regexec(label_pattern, spreadsheet_parameters))[[1]][1]
-      }
-      # Clean up to extract the date (removes name=as.Date( and the surrounding quotes)
-      label <- gsub(paste0(label_form, '"|"|\\)'), '', label_match)
-    } else if (list == FALSE){
-      if (logical == FALSE){
-        label_pattern <- paste0(label_form, '"(.*?)"')
-        label_match <- regmatches(spreadsheet_parameters, regexec(label_pattern, spreadsheet_parameters))[[1]][1]
-        label <- gsub(paste0(label_form, '"|"'), '', label_match)
-      } else {
-        label <- as.logical(trimws(gsub(paste0(".*", label_form), "", spreadsheet_parameters)))
-      }
-    } else {
-      label_pattern <- paste0(label_form, 'c\\((.*?)\\)')
-      label_match <- regmatches(spreadsheet_parameters, regexec(label_pattern, spreadsheet_parameters))[[1]][1]
-      label <- gsub(paste0(label_form), '', label_match)
-      label <- eval(parse(text = label))
-    }
-  } else {
-    label <- NULL
+  if (logical) {
+    logic_pattern <- paste0(name, "\\s*=\\s*(TRUE|FALSE)")
[…]
