### 2025-03-26 581c1fa
tests in bar_table
