+++ tests/testthat/test-shiny-reactivity.R
@@ -194,3 +194,199 @@ test_that("render_tabbed_display_items works correctly", {
   
   expect_silent(render_tabbed_display_items(display_box, display_content, data_list, output))
 })
+
+
+
+################################################################################
+################################################################################
+################################################################################
+################################################################################
+################################################################################
+################################################################################
+
+# 📁 tests/testthat/test-build_server.R
+
+library(testthat)
+library(shiny)
+library(dplyr)
+library(mockery)
+
+# --- 🔧 Helpers ---
+
+# Minimal reactive placeholder
+fake_reactive <- function(value) shiny::reactive(value)
+
+# Minimal display contents
+fake_contents <- function(name = "Display", id = "display", type = "Display") {
+  data.frame(name = name,
+             type = type, 
+             ID = id,
+             icon = "users",
+             stringsAsFactors = FALSE)
+}
+
+# Minimal main page with group or filter
+fake_main_page <- function(type = c("group_by_box", "filter_box"), var = "cyl") {
+  data.frame(
+    name = paste0(type, 1),
+    type = type,
+    parameter_list = c("label = \"CYL\", value = FALSE", "label = \"CYL\", choices = c(\"4\", \"6\", \"8\"), selected = c(\"4\", \"6\")"),
+    variable = var,
+    data = NA,
+    stringsAsFactors = FALSE
+  )
+}
+
+fake_demographics_page <- function(){
+  data.frame(
+    name = "box1",
+    type = "box",
+    value = "bar_table",
+    parameter_list = c("text = \"Gender\", colour = \"blue\""),
+    variable = "cyl",
+    row = 1,
+    data = NA
+  )
+}
+
+# --- 🔍 Tests ---
+
+test_that("build_server runs when data_list has no main_page", {
+  data_list <- list(contents = fake_contents())
+  df <- mtcars
+  
+  server_fn <- build_server(data_list, df, key_var = NULL, data_frame_name = "df")
+  
+  testServer(server_fn, {
+    expect_type(session, "environment")
+  })
+})
+
+test_that("build_server handles Tabbed_display types", {
+  data_list <- list(
+    main_page = NULL,
+    contents = rbind(
+      fake_contents("tab1", "Tabbed_display"),
+      fake_contents("tab2", "Display")
+    )
+  )
+  df <- mtcars
+  
+  server_fn <- build_server(data_list, df, key_var = NULL, data_frame_name = "df")
+  
+  testServer(server_fn, {
+    session$setInputs(tab = "tab1")
+    expect_true(TRUE)
+  })
+})
+
+test_that("build_server handles group_by_box logic", {
+  data_list <- list(
+    main_page = fake_main_page("group_by_box"),
+    contents = fake_contents(),
+    display = fake_demographics_page()
+  )
+  df <- mtcars
+  df$cyl <- factor(df$cyl)
+  
+  server_fn <- build_server(data_list, df, key_var = NULL, data_frame_name = "df")
+  
+  testServer(server_fn, {
+    session$setInputs(goButton_group = 1, group_by_box1 = TRUE)
+    expect_true(TRUE)  # Placeholder for coverage
+  })
+})
+
+test_that("build_server handles filter_box with variable only", {
+  data_list <- list(
+    main_page = fake_main_page("filter_box"),
+    contents = fake_contents(),
+    display = fake_demographics_page()
+  )
+  df <- mtcars
+  
[…]
