+++ tests/testthat/test-shiny-reactivity.R
@@ -363,22 +363,21 @@ test_that("build_server sets up download UI with credentials", {
   spreadsheet <- data.frame(
     type = c("Credentials", "Download label", "Format", "Data"),
     name = c("credentials", "Download", NA, "mtcars"),
-    value = c(NA, NA, "csv", "NHANES"),
+    value = c(NA, NA, "csv", "mtcars"),
     stringsAsFactors = FALSE
   )
   data_list <- list(
     contents = contents_df,
     download_sheet = spreadsheet
   )
-  df <- mtcars
-  
+
   credentials_data <- data.frame(
     user = "admin",
     password = "password",
     stringsAsFactors = FALSE
   )
   
-  server_fn <- build_server(data_list, df, key_var = "row.names", data_frame_name = "df")
+  server_fn <- build_server(data_list, mtcars, key_var = "row.names", data_frame_name = "mtcars")
   
   testServer(server_fn, {
     expect_type(output$build_download1, "list")
