### 2025-03-31 IDEMSInternational/ExcelToShiny#13 @devarops
Hi @lilyclements,

JOSS does not require a specific level of test coverage, but it does require that the functionality of the software can be verified through automated tests. So while high coverage isn't mandatory, having some meaningful tests in place is important for the review.

I’d encourage you to add tests, and fix the coverage report, rather than removing it. This will help demonstrate the reliability and correctness of the software.
