### 2025-05-13 IDEMSInternational/ExcelToShiny#44 @devarops
[JOSS REVIEW] Fix commented lines in installation code
Could you please remove the pound signs (#) from the installation code block? This will ensure the commands execute properly.

Instead of:
```r
# install.packages("devtools")
# devtools::install_github("IDEMSInternational/ExcelToShiny")
```

It should be:
```r
install.packages("devtools")
devtools::install_github("IDEMSInternational/ExcelToShiny")
```

Thanks!

---

This issue is part of the review process for a submission to the Journal of Open Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
