### 2025-05-13 IDEMSInternational/ExcelToShiny#45 @devarops
[JOSS REVIEW] Update the example in the "Example Usage" section of the README file
The example provided in the "Example Usage" section does not seem to work as intended. When I run the command:

```r
# Your shiny app
build_shiny(
  title = "Test Dashboard",
  data_list = example_excel,
  data_frame = NHANES,
  key_var = "ID"
)
```

I receive the following error:

```r
Error: object 'example_excel' not found
```

I tried replacing `example_excel` with `all_sheets`, which allowed the Shiny app to run. However, I am not sure if this substitution is correct or aligns with the intended usage. Could you clarify what the `data_list` argument should reference in this context?

Could you please update the example in the "Example Usage" section so it runs as intended?

![Image](https://github.com/user-attachments/assets/09433cb5-83c4-4f21-bea4-d0960f14b348)

![Image](https://github.com/user-attachments/assets/2a83f35d-8a5e-4d7c-a431-b523f5a7dd28)

---

This issue is part of the review process for a submission to the Journal of Open Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
