### 2025-05-13 IDEMSInternational/ExcelToShiny#46 @devarops
[JOSS REVIEW] Update example in `build_shiny()` documentation
The example in the documentation for the build_shiny() function appears to be outdated. When I attempt to run the provided code:

```r
# Run the dashboard
shinyApp(ui = PLH_shiny("My Dashboard", data_list, data_frame), server = function(input, output) {})
```

I encounter the following error:

```r
Error in PLH_shiny("My Dashboard", data_list, data_frame) :
  could not find function "PLH_shiny"
```

It seems `PLH_shiny()` is referenced instead of `build_shiny()`.

Could you please update the documentation example to use the correct function name and ensure it runs as intended?

---

This issue is part of the review process for a submission to the Journal of Open Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
