### 2025-05-13 IDEMSInternational/ExcelToShiny#47 @devarops
[JOSS REVIEW] Add minimal reproducible example for `generate_data_list()`
Could you please add a minimal reproducible example to the documentation for the `generate_data_list()` function?

Currently, the example assumes that `ODK_sample_survey` and `ODK_sample_data` are preloaded, which may cause confusion for users who don’t have these datasets available. Including a self-contained example (e.g., with code to generate sample data or use built-in datasets) would ensure clarity and reproducibility.

Thank you.

---

This issue is part of the review process for a submission to the Journal of Open Source Software (JOSS):

- https://github.com/openjournals/joss-reviews/issues/7876
