+++ R/generate_from_ODK_functions.R
@@ -85,12 +85,6 @@ create_display_pages_from_metadata_ODK <- function(input_data = ODK_sample_surve
 #'   - `contents`: A data frame with `name`, `type`, `ID`, and `icon` for each displayable group.
 #'   - `main_page`: A hardcoded main page data frame.
 #'   - Group-specific display tabs generated by `create_display_pages_from_metadata`.
-#' @export
-#'
-#' @examples
-#' # Assuming `ODK_sample_survey` and `ODK_sample_data` are loaded:
-#' # result <- generate_data_list(input_data = ODK_sample_survey, output_data = ODK_sample_data)
-#'
 generate_data_list <- function(input_data, output_data) {
   
   # Step 1: Filter to variables that exist in both the form and the data
