### 2025-05-14 97b54a5
Updating example in `build_shiny()` function
