### 2025-05-14 IDEMSInternational/ExcelToShiny#45 @lilyclements
@devarops thanks! I've updated the example, and this now runs as expected. You're absolutely right that it should be `all_sheets`, not `example_excel`.
