### 2025-05-14 IDEMSInternational/ExcelToShiny#45 @lilyclements
Fixed in #50
