### 2025-05-22 8b3e9c8
update and fixes in the download functionality
