### 2025-05-22 bb8072d
Dynamic option for viewing table
