+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: ExcelToShiny
 Title: Automate Shiny Dashboard Creation from Excel Templates
-Version: 0.1.0
+Version: 0.1.1
 Authors@R: 
     person("Lily", "Clements", , "lily@idems.international", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0001-8864-0552"))
@@ -27,7 +27,8 @@ Imports:
   stringdist,
   tibble,
   tidyr,
-  viridis
+  viridis,
+  writexl
 Suggests: 
     callr,
     devtools,
