### 2025-07-09 #7876 @lilyclements
@skanwal I have completed this list.

* I have made the release, and the version number is 0.2.0
* [DOI: 10.5281/zenodo.15846858](https://doi.org/10.5281/zenodo.15846858)
