### 2025-07-16 #7876 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
