### 2025-07-18 #7876 @crvernon
:wave: @lilyclements - your submission is looking great.  I just need you to address the following in your paper before I accept:

- [ ] LINE 101:  The "s" in "shiny" should be capitalized.  You can do this in your bib file by wrapping characters that you want to maintain the formatting of in curly brackets.
- [ ] LINE 107:  The "r" and "s" should be capitalized in "r shiny".

That's it.  No new release required.  Let me know when the changes have been made.
