### 2025-07-18 8bb3f7e
bibliography fix in capitalisation
