+++ joss_papers/paper.bib
@@ -26,7 +26,7 @@
 
 @misc{konrad2020shiny,
   author = {Markus Konrad},
-  title = {Developing a complex R Shiny app – the good, the bad and the ugly},
+  title = {Developing a complex {R} {S}hiny app – the good, the bad and the ugly},
   howpublished = {\url{https://datascience.blog.wzb.eu/2020/02/27/developing-a-complex-r-shiny-app/}},
   year = {2020}
 }
@@ -40,7 +40,7 @@
 
 @misc{biswas2021tableau,
   author = {Debolina Biswas},
-  title = {Tableau Vs Shiny: Which One Should You Pick For Data Visualisation},
+  title = {Tableau Vs {S}hiny: Which One Should You Pick For Data Visualisation},
   howpublished = {\url{https://analyticsindiamag.com/ai-origins-evolution/tableau-vs-shiny-which-one-should-you-pick-for-data-visualisation/}},
   year = {2021},
   month = oct,
